// client/src/lib/api/generated/types/AppointmentOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { AppointmentDetailsOut } from './AppointmentDetailsOut'

export type AppointmentOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    shop_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    barber_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    service_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    customer_user_id: string;
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    start_at: string;
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    end_at: string;
    status: string;
    booking_note: (string | null);
    cancelled_at: (string | null);
    completed_at: (string | null);
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    created_at: string;
    details?: (AppointmentDetailsOut | null);
};