// client/src/lib/api/generated/types/BarberManageOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type BarberManageOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    shop_member_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    user_id: string;
    display_name: string;
    bio: (string | null);
    profile_image_url: (string | null);
    status: string;
};