// client/src/lib/api/generated/types/BarberPublicOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { BarberServicePublicOut } from './BarberServicePublicOut'

export type BarberPublicOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    shop_id: string;
    name: string;
    bio: (string | null);
    profile_image_url: (string | null);
    rating: number;
    review_count: number;
    services: BarberServicePublicOut[];
};