// client/src/lib/api/generated/types/ReviewOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ReviewOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    appointment_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    shop_id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    customer_user_id: string;
    barber_profile_id: (string | null);
    rating: number;
    review_text: (string | null);
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    created_at: string;
};