// client/src/lib/api/generated/types/ServiceManageOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ServiceManageOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    shop_id: string;
    name: string;
    description: (string | null);
    /**
     * @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
     * @type string
    */
    price: string;
    currency: string;
    status: string;
};