// client/src/lib/api/generated/types/ServicePublicOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ServicePublicOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    name: string;
    description: (string | null);
    /**
     * @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
     * @type string
    */
    price: string;
    currency: string;
};