// client/src/lib/api/generated/types/ShopCustomerSummaryOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ShopCustomerSummaryOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    customer_user_id: string;
    display_name: (string | null);
    email: (string | null);
    phone: (string | null);
    visits: number;
    last_visit: (string | null);
    preferred_barber_id: (string | null);
};