// client/src/lib/api/generated/types/ShopPublicOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { OpeningHoursOut } from './OpeningHoursOut'

export type ShopPublicOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    name: string;
    tagline: string;
    description: string;
    photos: string[];
    rating: number;
    review_count: number;
    distance_km: (number | null);
    area: string;
    address: string;
    phone: (string | null);
    timezone: string;
    status: string;
    hours: OpeningHoursOut[];
};